+++ tests/unique_test.ts
@@ -1,5 +1,7 @@
+import { assertEquals } from "https://deno.land/std@0.140.0/testing/asserts.ts";
+import { tee } from "../tee.ts";
 import { assertStreams } from "../testing.ts";
-import { unique } from "../unique.ts";
+import { groupBy, unique } from "../unique.ts";
 import { getAsyncIterable } from "./testing_test.ts";
 
 Deno.test("unqiue(source)", async () => {
@@ -35,4 +37,27 @@ Deno.test("unique(source, key)", async () => {
     ),
     ["foo", "quux"],
   );
+});
+
+Deno.test(`groupBy()`, async () => {
+  const iterable = getAsyncIterable<{ id: number; name: string }>(
+    { id: 1, name: "foo" },
+    { id: 2, name: "bar" },
+    { id: 3, name: "bar" },
+    { id: 4, name: "foo" },
+  );
+  const [sourceA, sourceB] = tee(iterable, 2);
+  const expected = new Map();
+  expected.set("foo", [
+    { id: 1, name: "foo" },
+    { id: 4, name: "foo" },
+  ]);
+  expected.set("bar", [
+    { id: 2, name: "bar" },
+    { id: 3, name: "bar" },
+  ]);
+  let actual = await groupBy(sourceA, (o) => o.name);
+  assertEquals(actual, expected);
+  actual = await groupBy(sourceB, (o) => Promise.resolve(o.name));
+  assertEquals(actual, expected);
 });